requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pillow==11.3.0
numpy==1.26.4
google-generativeai
sentence-transformers
langextract

# LlamaIndex packages - let pip resolve a compatible set
lama-index
lama-parse
lama-index-embeddings-huggingface
lama-index-llms-gemini
